+++ app/src/main/java/com/st0rmtroop3r/telegramchart/enitity/ChartLine.java
@@ -8,6 +8,6 @@ public class ChartLine {
 
     public String color;
 
-    public long[] yData;
+    public int[] yData;
 
 }
